+++ src/dreamstate/core/observing/shouldObserversUpdate.ts
@@ -2,8 +2,11 @@ import { shallowEqualObjects } from "shallow-equal";
 
 import { dev } from "@/macroses/dev.macro";
 
+import { ComputedValue } from "@/dreamstate/core/observing/ComputedValue";
 import { NestedStore } from "@/dreamstate/core/observing/NestedStore";
 import { IStringIndexed } from "@/dreamstate/types";
+
+// todo: Warn computed/nested switch between states when previous and next types are different?
 
 /**
  * Compare context manager state diff with shallow check + nested objects check.
@@ -18,9 +21,20 @@ export function shouldObserversUpdate<T extends object>(
     }
   }
 
+  if (!previousContext) {
+    return true;
+  }
+
   return (
-    !previousContext ||
     Object.keys(nextContext).some(function(key: string): boolean {
+      /**
+       * Ignore computed values.
+       * Since nested computed stores are not representing data itself, we should not verify anything there.
+       */
+      if (nextContext[key] instanceof ComputedValue) {
+        return false;
+      }
+
       /**
        * Shallow check for mutable objects created by library.
        * Optimization for sub-states to prevent pollution of context and improve performance.