+++ src/dreamstate/types/utils.ts
@@ -13,8 +13,18 @@ export interface ILoadable<T, E = Error> extends INestedStore {
   asUpdated(value: T): ILoadable<T, E>;
 }
 
-export interface IMutable<T> extends INestedStore {
-  asMerged(state: Partial<T>): TMutable<T>;
+export type TLoadable<T, E = Error> = ILoadable<T, E>;
+
+export interface INested<T> extends INestedStore {
+  asMerged(state: Partial<T>): TNested<T>;
 }
 
-export type TMutable<T> = T & IMutable<T>;
+export type TNested<T> = T & INested<T>;
+
+export interface IComputed<T extends object, C extends object> {
+  __selector__: (context: C) => T;
+  __memo__?: (context: C) => Array<any>;
+  __diff__?: Array<any>;
+}
+
+export type TComputed<T extends object, C extends object = any> = T & IComputed<T, C>;